requirements.txt
python-docx
mtranslate
google_trans_new
googletrans